+++ .gitattributes
@@ -1,0 +1,2 @@
+build/* binary
+build/* linguist-generated=true
