+++ examples/preact/.prettierrc
