+++ examples/preact/capture-button.js
