+++ examples/preact/index-fallback.js
