+++ examples/preact/ops.js
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA
  */
 
-import initModule from '../libapi.mjs';
+import initModule from './build/libapi.mjs';
 
 /** @typedef {import('../libapi.mjs').Context} Context */
 
@@ -25,7 +25,8 @@ const ModulePromise = initModule();
 export function rethrowIfCritical(err) {
   // If it's precisely Error, it's a custom error; anything else - SyntaxError,
   // WebAssembly.RuntimeError, TypeError, etc. - is treated as critical here.
-  if (err.constructor !== Error) {
+  if (err.constructor !== Error)
+  {
     throw err;
   }
 }
@@ -44,7 +45,8 @@ export async function connect() {
    * @param {(ctx: Context) => Promise<T>} op
    * @returns {Promise<T>}
    */
-  function schedule(op) {
+  function schedule(op)
+  {
     let res = queue.then(() => op(context));
     queue = res.catch(rethrowIfCritical);
     return res;
