+++ examples/preact/serve.json
