+++ package.json
@@ -2,8 +2,8 @@
   "name": "web-gphoto2",
   "version": "0.1.0",
   "description": "WebAssembly implementation of gphoto2 and libusb to control DSLR cameras over USB on the Web",
-  "main": "ui/camera.js",
-  "types": "ui/libapi.mjs.d.ts",
+  "main": "src/camera.js",
+  "types": "src/libapi.mjs.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -23,8 +23,8 @@
   },
   "homepage": "https://github.com/GoogleChromeLabs/web-gphoto2#readme",
   "files": [
-    "ui/libapi.*",
-    "ui/camera.js"
+    "src/*",
+    "build/*"
   ],
   "devDependencies": {
     "@types/emscripten": "^1.39.5",
